+++ app/src/cli/mod.rs
@@ -1,3 +1,4 @@
 pub mod state;
 pub mod io;
+mod states;
 mod test_mock;
