+++ app/src/cli/states/exit.rs
@@ -1,0 +1,4 @@
+use super::CliState;
+
+pub struct Exit {}
+impl CliState<Exit> {}
