+++ app/src/cli/test_mock.rs
@@ -2,6 +2,6 @@ use std::io::Cursor;
 
 // 標準入力する値を受け取り、
 // その入力値が入力された状態の mock を生成する
-pub fn stdin_mock_with_inputted_text(input: &str) -> Cursor<&str> {
+pub fn _stdin_mock_with_inputted_text(input: &str) -> Cursor<&str> {
     Cursor::new(input)
 }
